pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "game-notifier"
version = "0.1.0"
description = "Generic tray notifier for game/content update feeds"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests",
  "Pillow",
  "pystray",
  "win10toast; sys_platform == 'win32'",
]

[project.gui-scripts]
game-notifier = "game_notifier.app:run"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = ["pytest>=7.0"]
